Make the changes in the right place
