+++ Assets/TinyMCE/JoplinLists/src/main/ts/api/Commands.ts
@@ -7,7 +7,8 @@
 
 import * as ToggleList from '../actions/ToggleList';
 import { indentListSelection, outdentListSelection, flattenListSelection } from '../actions/Indendation';
-import { addJoplinChecklistCommands } from '../listModel/JoplinListUtil';
+import { addJoplinChecklistCommands, isJoplinChecklistItem } from '../listModel/JoplinListUtil';
+
 
 const queryListCommandState = function (editor, listName) {
   return function () {
@@ -48,6 +49,26 @@ const register = function (editor) {
   editor.addQueryStateHandler('InsertDefinitionList', queryListCommandState(editor, 'DL'));
 
   addJoplinChecklistCommands(editor, ToggleList);
+
+  const editorClickHandler = (event) => {
+    if (!isJoplinChecklistItem(event.target)) return;
+
+    // We only process the click if it's within the checkbox itself (and not the label).
+    // That checkbox, based on
+    // the current styling is in the negative margin, so offsetX is negative when clicking
+    // on the checkbox itself, and positive when clicking on the label. This is strongly
+    // dependent on how the checkbox is styled, so if the style is changed, this might need
+    // to be updated too.
+    // For the styling, see:
+    // packages/renderer/MdToHtml/rules/checkbox.ts
+    //
+    // The previous solution was to use "pointer-event: none", which mostly work, however
+    // it means that links are no longer clickable when they are within the checkbox label.
+    if (event.offsetX >= 0) return;
+
+    editor.execCommand('ToggleJoplinChecklistItem', false, { element: event.target });
+  }
+  editor.on('click', editorClickHandler);
 };
 
 export {
